still working on paging, it's at least set up but no PF handler yet

+++ src/kernel/memory.c
@@ -7,10 +7,10 @@ dst - destination mem location
 src - source mem location
 num - number of bytes to copy
 */
-void* memcpy(void* dst, const void* src, uint16_t num) {
+void* memcpy(void* dst, const void* src, uint32_t num) {
     uint8_t* u8Dst = (uint8_t*) dst;
     const uint8_t* u8Src = (const uint8_t*) src;
-    for (uint16_t i = 0; i < num; ++i) {
+    for (uint32_t i = 0; i < num; ++i) {
         u8Dst[i] = u8Src[i];
     }
     return dst;
@@ -23,10 +23,10 @@ ptr - pointer to beginning of memory block
 value - value that will fill each byte of block
 num - bytes to fill
 */
-void* memset(void* ptr, int value, uint16_t num) {
+void* memset(void* ptr, int value, uint32_t num) {
     uint8_t* u8Ptr = (uint8_t*)ptr;
 
-    for(uint16_t i = 0; i < num; ++i) {
+    for(uint32_t i = 0; i < num; ++i) {
         u8Ptr[i] = (uint8_t)value;
     }
 
@@ -40,11 +40,11 @@ ptr1 - first memory block
 ptr2 - second memory block
 num - number of bytes to compare
 */
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num) {
+int memcmp(const void* ptr1, const void* ptr2, uint32_t num) {
     const uint8_t* u8Ptr1 = (const uint8_t*)ptr1;
     const uint8_t* u8Ptr2 = (const uint8_t*)ptr2;
 
-    for (uint16_t i = 0; i < num; ++i) {
+    for (uint32_t i = 0; i < num; ++i) {
         if (u8Ptr1[i] != u8Ptr2[i]) {
             return 1;   // false
         }
